feat:체크 필터 map써서 합치기 리팩토링링
