+++ src/pages/receive.tsx
@@ -7,7 +7,6 @@ import writing from "@public/assets/icon_writing.png";
 import mobilefilter from "@public/assets/icon_filterbutton.png";
 import ModalFilter from "@/components/Receive/ModalFilter";
 import { useState } from "react";
-import CheckFilter2 from "@/components/Receive/CheckFilter2";
 
 export default function Receive() {
   const [isOpen, setIsOpen] = useState<boolean>(false);
@@ -15,7 +14,7 @@ export default function Receive() {
     <div className="mt-[88px]">
       <p className="font-semibold text-[24px] py-[32px] ">받은 요청</p>
       <div className="flex gap-[107px] ">
-        <CheckFilter2 />
+        <CheckFilter />
         <div className="w-full">
           <CustomerInput />
           <div className="flex justify-between items-center mt-[24px] mb-[32px]">
@@ -36,7 +35,9 @@ export default function Receive() {
             <RequestDetails />
             <div className="flex justify-between items-center gap-[11px] mobile-tablet:gap-[8px] mobile:flex-col mobile:justify-normal">
               <button className="flex justify-center items-center gap-[10px] bg-color-blue-300 rounded-[16px] w-[448px] h-[64px] mobile:w-[300px] mobile:h-[48px] tablet:w-[280px] tablet:h-[48px]">
-                <p className="text-white font-semibold text-[20px] leading-[32px] whitespace-nowrap">견적 보내기</p>
+                <p className="text-white font-semibold text-[20px] leading-[32px] whitespace-nowrap">
+                  견적 보내기
+                </p>
                 <Image src={writing} alt="send" width={24} height={24} />
               </button>
               <button className="bg-white rounded-[16px] border border-color-blue-300 w-[448px] h-[64px] text-color-blue-300 font-semibold text-[20px] leading-[32px] mobile:w-[300px] mobile:h-[48px] tablet:w-[280px] tablet:h-[48px]">
